[REF] core: odoo.orm package
Move the files so that history is followed inside `_orm`, but do not
recreate them and instead create simple folders.
Adapt all necessary code inside odoo.orm and expose it in places where
the code is coming from for backward-compatibility and still an easy API.
