+++ odoo/fields/__init__.py
@@ -1,0 +1,25 @@
+# ruff: noqa: F401
+# Exports features of the ORM to developers.
+# This is a `__init__.py` file to avoid merge conflicts on `odoo/fields.py`.
+
+from odoo.orm.fields import Field
+
+from odoo.orm.fields import Id, Json
+from odoo.orm.fields import Boolean
+from odoo.orm.fields import Integer, Float, Monetary
+from odoo.orm.fields import Char, Text, Html
+from odoo.orm.fields import Selection
+from odoo.orm.fields import Date, Datetime
+
+from odoo.orm.fields import Many2one, Many2many, One2many
+from odoo.orm.fields import Many2oneReference, Reference
+
+from odoo.orm.fields import Properties, PropertiesDefinition
+from odoo.orm.fields import Binary, Image
+
+from odoo.orm.fields import Command
+
+# TODO these should not be exposed here
+from odoo.orm.fields import determine, first, NO_ACCESS
+from datetime import date, datetime
+from odoo.tools import date_utils
