+++ odoo/modules/registry/__init__.py
@@ -1,0 +1,3 @@
+# ruff: noqa: F401
+# Exposed here so that exist code is unaffected.
+from odoo.orm.registry import Registry
